remove trajectory_ca folder and move everything to trajectory
