+++ include/steam.hpp
@@ -61,6 +61,6 @@
 
 // trajectory
 #include <steam/trajectory/SteamTrajInterface.hpp>
-#include <steam/trajectory_ca/SteamCATrajInterface.hpp>
+#include <steam/trajectory/SteamCATrajInterface.hpp>
 
 #endif // STEAM_ESTIMATION_LIBRARY_HPP
