+++ include/steam/trajectory/SteamCATrajVar.hpp
